+++ packages/data/src/dataknobs_data/__init__.py
@@ -120,7 +120,7 @@ from .query_logic import (
 from .records import Record
 from .dedup import DedupChecker, DedupConfig, DedupResult, SimilarItem
 from .keyed_store import AsyncKeyedRecordStore, SyncKeyedRecordStore
-from .streaming import StreamConfig, StreamProcessor, StreamResult
+from .streaming import ConflictPolicy, StreamConfig, StreamProcessor, StreamResult
 from .transactions import VALID_TRANSACTION_POLICIES, BufferedTransaction
 
 __version__ = "0.5.5"
@@ -157,6 +157,7 @@ __all__ = [
     "StreamConfig",
     "StreamResult",
     "StreamProcessor",
+    "ConflictPolicy",
     # Transactions
     "BufferedTransaction",
     "VALID_TRANSACTION_POLICIES",